+++ example/src/l18n/intl_config.ts
@@ -20,7 +20,14 @@ const intlConfig = setIntlConfig({
         redirectAuthPath: '/login',
         homePath: '/',
         isAuthPath: (path: string) => path === '/login',
-        verifyEmailPath: '/test'
+        verifyEmailPath: '/test',
+        appCheck: {
+            recaptchaV3SiteKey: process.env.NEXT_PUBLIC_RECAPTCHA_V3_SITE_KEY as string,
+            debugToken: process.env.NODE_ENV === "development" ? "f4bd569c-48e3-4351-87b6-e32466030271" : undefined,
+            privateKey: process.env.FIREBASE_SERVICE_ACCOUNT_PRIVATE_KEY!,
+            clientEmail: process.env.FIREBASE_SERVICE_ACCOUNT_CLIENT_EMAIL!,
+            appId: process.env.FIREBASE_APP_ID!,
+        },
     } : undefined,
     generate: {
         getCloudflareContext: getCloudflareContext,